Bug fix - Constructing Right camera object
Date: <19/02/2015>

Author: Anoop Vargheese
Email: anoopvargheese@gmail.com

Changelist:
* In src/video_input/MetricCamera.cpp, changed line 58 for allocation of Projection Matrix.

+ Also added .gitingore files for the /ground_estimation, /stixel_world, and /objects_detection folders to ignore Cmake and generated files

Extra Notes:
The hotfix for the ProjectionMatrix error is weird. The program keeps seg-faulting whenever it tries to set the projection matrix for the second camera. The error seems to be in the allocation (operator=) of the Eigen Matrix class. Found a workaround by using the TopLeftCorner() allocator instead.